A live thumbnail sometimes is not updated.
In ico-uxf-weston-plugin(weston plugin), when Weston will finish
a certain processing and will be in an idle state, it investigated
whether the buffer number of surface changed, and if it is changing,
it will have judged with drawing having changed.
A buffer number changes because an application calls SwapBuffer.
However, since Weston may process at once 2 times of SwapBuffers which
the application called continuously and a buffer number returns by
SwapBuffer in that case, drawing judges that it is not changing and
does not remake a live thumbnail.
Even if the buffer number of repair was the same, when having passed
beyond the definite period of time, it corresponded by remaking a live
thumbnail.

Change-Id: Id241741344a4b8e7736dc5efad06c5ece1d78810
Signed-off-by: Ryota Okubi <ryota.okubi@mail.toyota-td.jp>