The HomeScreen operation by the switch of a handle is not effective.
The menu of HomeScreen is not displayed even if it pushes the switch of
G27 driving handle.The event of this switch input is specification sent to
HomeScreen as a keyboard event.Acquisition processing of keyboard resource
of ico_input_mgr continues being Weston 1.4.0, and since a resource does
not exist for this reason, the cause cannot send an event.
Repair corresponded by changing acquisition processing of keyboard resource
according to Weson 1.5.0.
Moreover, since the argument of ivi_layout_surfaceSetDimension() of
GENIVI-layerManagement was changed into an int type from a uint type and
warning was outputted, it corrected.

Change-Id: I27ae1909d80a5f2ecb9f12e3cca45061b42e0d54
Signed-off-by: Ryota Okubi <ryota.okubi@mail.toyota-td.jp>